+++ client/packages/attester/config/roco.ts
@@ -1,11 +1,11 @@
 export default {
     circuit: {
         rpc1: {
-            ws: 'wss://ws.t0rn.io',
+            ws: 'wss://rpc.t0rn.io',
             http: 'https://rpc.t0rn.io',
         },
         rpc2: {
-            ws: 'wss://ws.t0rn.io',
+            ws: 'wss://rpc.t0rn.io',
             http: 'https://rpc.t0rn.io',
         },
     },