+++ client/packages/executor/src/circuit/listener.ts
@@ -60,6 +60,8 @@ export class CircuitListener extends EventEmitter {
 
   async start() {
     this.stop = await this.client.query.system.events((notifications) => {
+      // TODO: we should also monitor what events we are receiving here
+      // TODO: refactor this to use event types in the same way ExecutionManager does
       for (let i = 0; i < notifications.length; i++) {
         if (notifications[i].event.method === "NewSideEffectsAvailable") {
           // receives new side effects