+++ client/packages/executor/src/gateways/substrate/relayer.ts
@@ -61,6 +61,7 @@ export class SubstrateRelayer extends EventEmitter {
         nonce: this.nonce,
         target: this.name,
         balance: balance,
+        rpc: config.rpc,
       },
       "Relayer setup completed 🏁",
     );
@@ -299,6 +300,21 @@ export class SubstrateRelayer extends EventEmitter {
     const blockHash = await this.client.rpc.chain.getBlockHash(
       relaychainBlockNumber,
     );
+    logger.debug(
+      {
+        blockHash,
+        relaychainBlockNumber,
+        parachainId,
+        inclusionProof: Utils.Substrate.getStorageProof(
+          this.client,
+          blockHash,
+          "Paras",
+          "Heads",
+          parachainId,
+        ),
+      },
+      "Generating header inclusion proof with Substrate.getStorageProof",
+    );
 
     return Utils.Substrate.getStorageProof(
       this.client,
@@ -362,7 +378,7 @@ export class SubstrateRelayer extends EventEmitter {
   async fetchNonce(api: ApiPromise, address: string): Promise<number> {
     return await api.rpc.system.accountNextIndex(address).then((nextIndex) => {
       // @ts-ignore - property does not exist on type
-      return parseInt(nextIndex.toHuman());
+      return parseInt(nextIndex.toNumber());
     });
   }
 }